App: Main-menu: Move main menu to a separate component
